Accessible Game State and Other misc (#194)
* Move gamestate to public namespace
* memAllocLocal
* Writable MapStringToClass
* game_variable
* Fix r_string cmp
* Create compactArray from iterators
* Option to ignore plugin cert in Debug build
* Proper Iterators for compact_array
* Change eventhandler ID's to be signed int **Plugin Breaking change**

+++ src/client/headers/client/eventhandlers.hpp
@@ -31,7 +31,7 @@ namespace intercept::client {
     /// @private
     struct EHIdentifier {
         int32_t internal_id;
-        float arma_eh_id;
+        int arma_eh_id;
         uint32_t EHIteration;
         uint8_t EHType;
         bool already_deleted = false;
